Revert "기능 7번 구현"
