--- .gitignore
@@ -1,12 +1,0 @@
-### Example user template template
-### Example user template
-
-# IntelliJ project files
-.idea
-*.iml
-out
-gen
-bin/
-.settings
-.classpath
-.project
